Use abstract Subscription and Stream types
