+++ crates/relayer-all-in-one/src/one_for_all/impls/chain/event_subscription.rs
@@ -1,6 +1,5 @@
-use alloc::sync::Arc;
-
 use ibc_relayer_components::chain::traits::event_subscription::HasEventSubscription;
+use ibc_relayer_components::runtime::traits::subscription::HasSubscriptionType;
 
 use crate::one_for_all::traits::chain::OfaChain;
 use crate::one_for_all::types::chain::OfaChainWrapper;
@@ -9,7 +8,9 @@ impl<Chain> HasEventSubscription for OfaChainWrapper<Chain>
 where
     Chain: OfaChain,
 {
-    fn event_subscription(&self) -> &Arc<dyn Subscription<Item = (Self::Height, Self::Event)>> {
+    fn event_subscription(
+        &self,
+    ) -> &<Self::Runtime as HasSubscriptionType>::Subscription<(Self::Height, Self::Event)> {
         self.chain.event_subscription()
     }
 }
