+++ crates/relayer-cosmos/src/contexts/chain.rs
@@ -5,8 +5,8 @@ use ibc_relayer::chain::handle::ChainHandle;
 use ibc_relayer::config::EventSourceMode;
 use ibc_relayer::event::source::queries::all as all_queries;
 use ibc_relayer::keyring::Secp256k1KeyPair;
-use ibc_relayer_components::runtime::impls::subscription::empty::EmptySubscription;
-use ibc_relayer_components::runtime::traits::subscription::Subscription;
+use ibc_relayer_runtime::impls::subscription::empty::EmptySubscription;
+use ibc_relayer_runtime::traits::subscription::Subscription;
 use ibc_relayer_runtime::types::runtime::TokioRuntimeContext;
 use ibc_relayer_types::core::ics02_client::height::Height;
 use ibc_relayer_types::core::ics24_host::identifier::ChainId;
