+++ crates/relayer-runtime/src/impls/runtime/channel.rs
@@ -102,7 +102,7 @@ impl CanUseChannelsOnce for TokioRuntimeContext {
 impl CanStreamReceiver for TokioRuntimeContext {
     fn receiver_to_stream<T>(
         receiver: Self::Receiver<T>,
-    ) -> Pin<Box<dyn Stream<Item = T> + Send + 'static>>
+    ) -> Pin<Box<dyn Stream<Item = T> + Send + Sync + 'static>>
     where
         T: Async,
     {
