+++ crates/relayer-runtime/src/impls/runtime/mod.rs
@@ -4,5 +4,6 @@ pub mod mutex;
 pub mod sleep;
 pub mod spawn;
 pub mod stream;
+pub mod subscription;
 pub mod task;
 pub mod time;
