+++ crates/relayer-runtime/src/impls/subscription/empty.rs
@@ -3,10 +3,9 @@ use core::pin::Pin;
 
 use async_trait::async_trait;
 use cgp_core::traits::Async;
-use futures_core::stream::Stream;
+use futures::stream::Stream;
 
-use crate::runtime::traits::subscription::Subscription;
-use crate::std_prelude::*;
+use crate::traits::subscription::Subscription;
 
 pub struct EmptySubscription<T>(pub PhantomData<T>);
 
@@ -26,7 +25,9 @@ impl<T> EmptySubscription<T> {
 impl<T: Async> Subscription for EmptySubscription<T> {
     type Item = T;
 
-    async fn subscribe(&self) -> Option<Pin<Box<dyn Stream<Item = Self::Item> + Send + 'static>>> {
+    async fn subscribe(
+        &self,
+    ) -> Option<Pin<Box<dyn Stream<Item = Self::Item> + Send + Sync + 'static>>> {
         None
     }
 }
