+++ crates/relayer-runtime/src/impls/subscription/mod.rs
@@ -1,2 +1,4 @@
 pub mod closure;
 pub mod empty;
+pub mod multiplex;
+pub mod stream;
